refactor: migrate ThreeViewerService
ref #2318
