+++ visualization/app/app.module.ts
@@ -15,11 +15,7 @@ import { EffectsModule } from "./codeCharta/state/angular-redux/effects/effects.
 import { UnfocusNodesOnLoadingMapEffect } from "./codeCharta/state/effects/unfocusNodesOnLoadingMap/unfocusNodesOnLoadingMap.effect"
 import { AddBlacklistItemsIfNotResultsInEmptyMapEffect } from "./codeCharta/state/effects/addBlacklistItemsIfNotResultsInEmptyMap/addBlacklistItemsIfNotResultsInEmptyMap.effect"
 import { dialogs } from "./codeCharta/ui/dialogs/dialogs"
-import {
-	codeChartaServiceProvider,
-	CodeMapMouseEventServiceTokenProvider,
-	threeViewerServiceTokenProvider
-} from "./codeCharta/services/ajs-upgraded-providers"
+import { codeChartaServiceProvider, CodeMapMouseEventServiceTokenProvider } from "./codeCharta/services/ajs-upgraded-providers"
 import { NodeContextMenuCardModule } from "./codeCharta/state/effects/nodeContextMenu/nodeContextMenuCard/nodeContextMenuCard.module"
 import { OpenNodeContextMenuEffect } from "./codeCharta/state/effects/nodeContextMenu/openNodeContextMenu.effect"
 import { FocusButtonsComponent } from "./codeCharta/state/effects/nodeContextMenu/focusButtons/focusButtons.component"
@@ -86,7 +82,6 @@ import { UpdateAttributeTypesEffect } from "./codeCharta/state/effects/updateAtt
 	],
 	providers: [
 		codeChartaServiceProvider,
-		threeViewerServiceTokenProvider,
 		CodeMapMouseEventServiceTokenProvider,
 		VersionService,
 		{
