+++ visualization/app/codeCharta/services/ajs-upgraded-providers.ts
@@ -9,15 +9,6 @@ export const codeChartaServiceProvider = {
 	deps: ["$injector"]
 }
 
-export const ThreeViewerServiceToken = new InjectionToken("ThreeViewerService")
-export const threeViewerServiceTokenProvider = {
-	provide: ThreeViewerServiceToken,
-	useFactory: function ThreeViewerServiceTokenFactory(injector: Injector) {
-		return injector.get("threeViewerService")
-	},
-	deps: ["$injector"]
-}
-
 export const CodeMapMouseEventServiceToken = new InjectionToken("CodeMapMouseEventService")
 export const CodeMapMouseEventServiceTokenProvider = {
 	provide: CodeMapMouseEventServiceToken,
