+++ visualization/app/codeCharta/ui/codeMap/threeViewer/threeViewer.module.ts
@@ -1,10 +1,9 @@
-import { ThreeViewerService } from "./threeViewerService"
+import { ThreeViewerService } from "./threeViewer.service"
 import { ThreeSceneService } from "./threeSceneService"
 import { ThreeCameraService } from "./threeCamera.service"
 import { ThreeOrbitControlsService } from "./threeOrbitControls.service"
 import { ThreeStatsService } from "./threeStats.service"
 import "../../../state/state.module"
-import camelCase from "lodash.camelcase"
 
 import angular from "angular"
 import { downgradeInjectable } from "@angular/upgrade/static"
@@ -13,7 +12,7 @@ import { ThreeRendererService } from "./threeRenderer.service"
 
 angular
 	.module("app.codeCharta.ui.codeMap.threeViewer", ["app.codeCharta.state"])
-	.service(camelCase(ThreeViewerService.name), ThreeViewerService)
+	.factory("threeViewerService", downgradeInjectable(ThreeViewerService))
 	.factory("threeCameraService", downgradeInjectable(ThreeCameraService))
 	.factory("threeSceneService", downgradeInjectable(ThreeSceneService))
 	.factory("threeStatsService", downgradeInjectable(ThreeStatsService))
