+++ visualization/app/codeCharta/ui/codeMap/threeViewer/threeViewer.service.ts
@@ -1,21 +1,21 @@
+import { Inject, Injectable } from "@angular/core"
 import { ThreeSceneService } from "./threeSceneService"
 import { ThreeCameraService } from "./threeCamera.service"
 import { ThreeOrbitControlsService } from "./threeOrbitControls.service"
 import { ThreeRendererService } from "./threeRenderer.service"
 import { ThreeStatsService } from "./threeStats.service"
 
+@Injectable({ providedIn: "root" })
 export class ThreeViewerService {
 	private animationFrameId: number
 
 	constructor(
-		private threeSceneService: ThreeSceneService,
-		private threeCameraService: ThreeCameraService,
-		private threeOrbitControlsService: ThreeOrbitControlsService,
-		private threeRendererService: ThreeRendererService,
-		private threeStatsService: ThreeStatsService
-	) {
-		"ngInject"
-	}
+		@Inject(ThreeSceneService) private threeSceneService: ThreeSceneService,
+		@Inject(ThreeCameraService) private threeCameraService: ThreeCameraService,
+		@Inject(ThreeOrbitControlsService) private threeOrbitControlsService: ThreeOrbitControlsService,
+		@Inject(ThreeRendererService) private threeRendererService: ThreeRendererService,
+		@Inject(ThreeStatsService) private threeStatsService: ThreeStatsService
+	) {}
 
 	init(target: Element) {
 		this.threeCameraService.init(window.innerWidth, window.innerHeight)
